big dqn

+++ src/main.py
@@ -1,5 +1,6 @@
 import random
 import os
+from pathlib import Path
 import numpy as np
 import torch
 
@@ -19,12 +20,14 @@ def seed_everything(seed: int = 42):
 
 
 if __name__ == "__main__":
-    seed_everything(seed=42)
-    # Initialization of the agent. Replace DummyAgent with your custom agent implementation.
-    agent = ProjectAgent()
-    agent.load()
-    # Keep the following lines to evaluate your agent unchanged.
-    score_agent: float = evaluate_HIV(agent=agent, nb_episode=1)
-    score_agent_dr: float = evaluate_HIV_population(agent=agent, nb_episode=15)
-    with open(file="score.txt", mode="w") as f:
-        f.write(f"{score_agent}\n{score_agent_dr}")
+    file = Path("score.txt")
+    if not file.is_file():
+        seed_everything(seed=42)
+        # Initialization of the agent. Replace DummyAgent with your custom agent implementation.
+        agent = ProjectAgent()
+        agent.load()
+        # Evaluate agent and write score.
+        score_agent: float = evaluate_HIV(agent=agent, nb_episode=5)
+        score_agent_dr: float = evaluate_HIV_population(agent=agent, nb_episode=20)
+        with open(file="score.txt", mode="w") as f:
+            f.write(f"{score_agent}\n{score_agent_dr}")
